Install Redux dependencites
TL;DR: Redux will be used for state management of this app.

The longer version:
In planning the app, I initially considered using React Context API, because it's light-weight and also newer & fun in combination React hooks. A nice opportunity to try our something new...
But for this app, we don't have deeply nested components. We have data coming from multiple API endpoints that influences a single view. We want to be able to monitor our data easily.
For these reasons, I decided that Redux was a nicer fit. Actions making API calls and dispatching the response to reducers is clean & easy to understand. Having access to Redux Dev Tools to monitor state is also nice.
So, perhaps I will save playing with Context API for another time!